updates to handle more cases
